move logic to re-usable class
